avoid fd limit issues in tests
On OS X, this is more pronounced because the default file descriptor
limit is 256 and the same temporary directory is used everywhere.

When justrun walks up to the parent directory for the detection of
renames in a child path below, it can run into the very large temporary
directories, watch the many children paths inside, and blow out the file
descriptor limit.

This makes a subdirectory for the sole person of avoiding that walk to
the parent blowing up. Instead, the parent will be the
justrun_tests_${NUM} directory, which is empty for each test.