clean up of reload, waitForInterrupt, and watch
reload no longer does weird things with a pointer to lastStartTime and,
instead, just returns the right time.

removed a defer in waitForInterrupt that wasn't as helpful as it had
been expected to be

adding a bunch of comments in watch and the Ignorers, and cleaning up
which loop through the user paths does what.