use syscall.ESRCH as exact error for shutdown
Previously, we would run a new command if the previous command was returning
"signal terminated". However, the process might be just in a soft, graceful
shutdown and still be cleaning up its resources when that was true.

Instead, look specifically for the process to disappear from the process
space. The done channel becomes a useful optimization that lets us avoid both
spin-looping to poll Terminate (if the time.After is removed), and waiting
longer than we need to for the process to be done.