cmdwrap: fix build on windows
Currently the build fails to compile on Windows machines because
syscall.Kill is not defined for Windows (reproduce with GOOS=windows
go install ./...). Use an alternative method to terminate the process
on Windows. This is not perfect but it's an improvement over the
current in that it compiles and does what we want.

The current best practice is to avoid directly using the `syscall`
package - the `x/sys/{unix,windows}` packages wrap syscall accesses
in build tags which provide slightly more safety than using
the syscall package directly; see the deprecation notice here.
https://golang.org/pkg/syscall/

Also use a *SysProcAttr on Windows that compiles - it doesn't have the
Setpgid attribute so that was failing to compile.

Noticed this while trying to compile and host a release of justrun on
go.equinox.io.