cmdwrap: omit erroneous "error in process termination" message
If you run justrun with the "-v" flag, and change a file, justrun will
terminate the supervised process by sending SIGTERM to it. However,
when the supervised process exits justrun would report that there was
an error in process termination, even if the error was that the
process was terminated:

2020/01/11 23:21:25 error in process termination of pid 47118: signal: terminated

Since we expect a terminated process to be terminated, reverse
engineer the error returned by the underlying *os.ProcessState and
omit the error log if that's the case.